requirements.txt
pytest>=6.2.4
pytest-playwright>=0.3.0
